// libs/schemas/src/audio/schemas.gen.ts
export const AceStepAudioInpaintOutputSchema = {
  type: "object",
  title: "ACEStepAudioInpaintResponse",
  required: ["audio", "seed", "tags", "lyrics"],
  properties: {
    tags: {
      description: "The genre tags used in the generation process.",
      type: "string",
      title: "Tags",
      examples: ["lofi, hiphop, drum and bass, trap, chill"],
    },
    audio: {
      description: "The generated audio file.",
      examples: [
        {
          url: "https://storage.googleapis.com/falserverless/example_outputs/ace-step-audio-inpaint.wav",
        },
      ],
      $ref: "#/$defs/File",
    },
    seed: {
      description: "The random seed used for the generation process.",
      type: "integer",
      title: "Seed",
      examples: [42],
    },
    lyrics: {
      description: "The lyrics used in the generation process.",
      type: "string",
      title: "Lyrics",
      examples: ["[inst]"],
    },
  },
  "x-fal-order-properties": ["audio", "seed", "tags", "lyrics"],
  $defs: {
    File: {
      title: "File",
      properties: {
        file_size: {
          examples: [4404019],
          title: "File Size",
          description: "The size of the file in bytes.",
          anyOf: [{ type: "integer" }, { type: "null" }],
        },
        file_name: {
          examples: ["z9RV14K95DvU.png"],
          title: "File Name",
          description:
            "The name of the file. It will be auto-generated if not provided.",
          anyOf: [{ type: "string" }, { type: "null" }],
        },
        content_type: {
          examples: ["image/png"],
          title: "Content Type",
          description: "The mime type of the file.",
          anyOf: [{ type: "string" }, { type: "null" }],
        },
        url: {
          title: "Url",
          type: "string",
          description: "The URL where the file can be downloaded from.",
        },
      },
      required: ["url"],
      type: "object",
      "x-fal-order-properties": [
        "url",
        "content_type",
        "file_name",
        "file_size",
      ],
    },
  },
} as const;

export const AceStepAudioOutpaintInputSchema = {
  type: "object",
  title: "ACEStepAudioOutpaintRequest",
  required: ["tags", "audio_url"],
  properties: {
    scheduler: {
      enum: ["euler", "heun"],
      description: "Scheduler to use for the generation process.",
      type: "string",
      title: "Scheduler",
      default: "euler",
      examples: ["euler"],
    },
    guidance_scale: {
      description: "Guidance scale for the generation.",
      minimum: 0,
      type: "number",
      title: "Guidance Scale",
      default: 15,
      maximum: 200,
      examples: [15],
    },
    granularity_scale: {
      description:
        "Granularity scale for the generation process. Higher values can reduce artifacts.",
      minimum: -100,
      type: "integer",
      title: "Granularity Scale",
      default: 10,
      maximum: 100,
      examples: [10],
    },
    tag_guidance_scale: {
      description: "Tag guidance scale for the generation.",
      minimum: 0,
      type: "number",
      title: "Tag Guidance Scale",
      default: 5,
      maximum: 10,
      examples: [5],
    },
    tags: {
      description:
        "Comma-separated list of genre tags to control the style of the generated audio.",
      type: "string",
      title: "Tags",
      examples: ["lofi, hiphop, drum and bass, trap, chill"],
    },
    seed: {
      description:
        "Random seed for reproducibility. If not provided, a random seed will be used.",
      title: "Seed",
      anyOf: [{ type: "integer" }, { type: "null" }],
    },
    guidance_interval: {
      description:
        "Guidance interval for the generation. 0.5 means only apply guidance in the middle steps (0.25 * infer_steps to 0.75 * infer_steps)",
      minimum: 0,
      type: "number",
      title: "Guidance Interval",
      default: 0.5,
      maximum: 1,
      examples: [0.5],
    },
    extend_after_duration: {
      description: "Duration in seconds to extend the audio from the end.",
      minimum: 0,
      maximum: 240,
      type: "number",
      examples: [30],
      title: "Extend After Duration",
      default: 30,
    },
    guidance_type: {
      enum: ["cfg", "apg", "cfg_star"],
      description: "Type of CFG to use for the generation process.",
      type: "string",
      title: "Guidance Type",
      default: "apg",
      examples: ["apg"],
    },
    minimum_guidance_scale: {
      description: "Minimum guidance scale for the generation after the decay.",
      minimum: 0,
      type: "number",
      title: "Minimum Guidance Scale",
      default: 3,
      maximum: 200,
      examples: [3],
    },
    number_of_steps: {
      description: "Number of steps to generate the audio.",
      minimum: 3,
      type: "integer",
      title: "Number Of Steps",
      default: 27,
      maximum: 60,
      examples: [27],
    },
    guidance_interval_decay: {
      description:
        "Guidance interval decay for the generation. Guidance scale will decay from guidance_scale to min_guidance_scale in the interval. 0.0 means no decay.",
      minimum: 0,
      type: "number",
      title: "Guidance Interval Decay",
      default: 0,
      maximum: 1,
      examples: [0],
    },
    lyrics: {
      description:
        "Lyrics to be sung in the audio. If not provided or if [inst] or [instrumental] is the content of this field, no lyrics will be sung. Use control structures like [verse], [chorus] and [bridge] to control the structure of the song.",
      type: "string",
      title: "Lyrics",
      default: "",
    },
    extend_before_duration: {
      description: "Duration in seconds to extend the audio from the start.",
      minimum: 0,
      maximum: 240,
      type: "number",
      examples: [0],
      title: "Extend Before Duration",
      default: 0,
    },
    lyric_guidance_scale: {
      description: "Lyric guidance scale for the generation.",
      minimum: 0,
      type: "number",
      title: "Lyric Guidance Scale",
      default: 1.5,
      maximum: 10,
      examples: [1.5],
    },
    audio_url: {
      description: "URL of the audio file to be outpainted.",
      type: "string",
      title: "Audio Url",
      examples: [
        "https://storage.googleapis.com/falserverless/example_inputs/ace-step-audio-to-audio.wav",
      ],
      "x-fal-file-input": true,
    },
  },
  "x-fal-order-properties": [
    "audio_url",
    "extend_before_duration",
    "extend_after_duration",
    "tags",
    "lyrics",
    "number_of_steps",
    "seed",
    "scheduler",
    "guidance_type",
    "granularity_scale",
    "guidance_interval",
    "guidance_interval_decay",
    "guidance_scale",
    "minimum_guidance_scale",
    "tag_guidance_scale",
    "lyric_guidance_scale",
  ],
} as const;

export const AceStepAudioOutpaintOutputSchema = {
  type: "object",
  title: "ACEStepResponse",
  required: ["audio", "seed", "tags", "lyrics"],
  properties: {
    tags: {
      description: "The genre tags used in the generation process.",
      type: "string",
      title: "Tags",
      examples: ["lofi, hiphop, drum and bass, trap, chill"],
    },
    audio: {
      description: "The generated audio file.",
      examples: [
        {
          url: "https://storage.googleapis.com/falserverless/example_outputs/ace-step-text-to-audio.wav",
        },
      ],
      $ref: "#/$defs/File",
    },
    seed: {
      description: "The random seed used for the generation process.",
      type: "integer",
      title: "Seed",
      examples: [42],
    },
    lyrics: {
      description: "The lyrics used in the generation process.",
      type: "string",
      title: "Lyrics",
      examples: ["[inst]"],
    },
  },
  "x-fal-order-properties": ["audio", "seed", "tags", "lyrics"],
  $defs: {
    File: {
      title: "File",
      properties: {
        file_size: {
          examples: [4404019],
          title: "File Size",
          description: "The size of the file in bytes.",
          anyOf: [{ type: "integer" }, { type: "null" }],
        },
        file_name: {
          examples: ["z9RV14K95DvU.png"],
          title: "File Name",
          description:
            "The name of the file. It will be auto-generated if not provided.",
          anyOf: [{ type: "string" }, { type: "null" }],
        },
        content_type: {
          examples: ["image/png"],
          title: "Content Type",
          description: "The mime type of the file.",
          anyOf: [{ type: "string" }, { type: "null" }],
        },
        url: {
          title: "Url",
          type: "string",
          description: "The URL where the file can be downloaded from.",
        },
      },
      required: ["url"],
      type: "object",
      "x-fal-order-properties": [
        "url",
        "content_type",
        "file_name",
        "file_size",
      ],
    },
  },
} as const;
// ...
export const AceStepAudioToAudioOutputSchema = {
  type: "object",
  title: "ACEStepAudioToAudioResponse",
  required: ["audio", "seed", "tags", "lyrics"],
  properties: {
    tags: {
      description: "The genre tags used in the generation process.",
      type: "string",
      title: "Tags",
      examples: ["lofi, hiphop, drum and bass, trap, chill"],
    },
    audio: {
      description: "The generated audio file.",
      examples: [
        {
          url: "https://storage.googleapis.com/falserverless/example_outputs/ace-step-audio-to-audio.wav",
        },
      ],
      $ref: "#/$defs/File",
    },
    seed: {
      description: "The random seed used for the generation process.",
      type: "integer",
      title: "Seed",
      examples: [42],
    },
    lyrics: {
      description: "The lyrics used in the generation process.",
      type: "string",
      title: "Lyrics",
      examples: ["[inst]"],
    },
  },
  "x-fal-order-properties": ["audio", "seed", "tags", "lyrics"],
  $defs: {
    File: {
      title: "File",
      properties: {
        file_size: {
          examples: [4404019],
          title: "File Size",
          description: "The size of the file in bytes.",
          anyOf: [{ type: "integer" }, { type: "null" }],
        },
        file_name: {
          examples: ["z9RV14K95DvU.png"],
          title: "File Name",
          description:
            "The name of the file. It will be auto-generated if not provided.",
          anyOf: [{ type: "string" }, { type: "null" }],
        },
        content_type: {
          examples: ["image/png"],
          title: "Content Type",
          description: "The mime type of the file.",
          anyOf: [{ type: "string" }, { type: "null" }],
        },
        url: {
          title: "Url",
          type: "string",
          description: "The URL where the file can be downloaded from.",
        },
      },
      required: ["url"],
      type: "object",
      "x-fal-order-properties": [
        "url",
        "content_type",
        "file_name",
        "file_size",
      ],
    },
  },
} as const;
// ...
export const AudioUnderstandingOutputSchema = {
  required: ["output"],
  properties: {
    output: {
      examples: [
        "Based on the audio, this appears to be a business meeting discussing quarterly sales results. The speakers are analyzing performance metrics and discussing strategies for the upcoming quarter. The tone is professional and collaborative, with multiple participants contributing to the discussion.",
      ],
      type: "string",
      title: "Output",
      description: "The analysis of the audio content based on the prompt",
    },
  },
  type: "object",
  "x-fal-order-properties": ["output"],
  title: "AudioUnderstandingOutput",
} as const;